Remove `extensions/components/groups` stale folder
