+++ features/admin.application-roles.v1/components/roles-list.tsx
@@ -18,7 +18,6 @@
 
 import { history } from "@wso2is/admin.core.v1";
 import { AppConstants, UIConstants } from "@wso2is/admin.core.v1/constants";
-import { RoleBasicInterface } from "@wso2is/admin.extensions.v1/components/groups/models";
 import { IdentifiableComponentInterface } from "@wso2is/core/models";
 import {
     DataTable,
@@ -30,6 +29,7 @@ import {
 import React, { ChangeEvent, ReactElement, ReactNode, SyntheticEvent, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { Grid, Header, Icon, Input, SemanticICONS } from "semantic-ui-react";
+import { RoleBasicInterface } from "../models/application-roles";
 
 interface RolesListProps extends IdentifiableComponentInterface {
     rolesList: RoleBasicInterface[];
