--- features/admin.extensions.v1/components/groups/data/group.md
@@ -1,9 +1,0 @@
-## Groups
-A collection of users who are granted the same privileges in the system. By grouping users, developers can use this group attribute information to write logic in their business applications.
-
-**Note:** Only users can be assigned to a group.
-
-### Source
-Where the groups' details are being managed. By default, there can be following sources.
-
-**Local/Consumer**: Default userstore where  groups' details are being managed.
